// src/app/core/api/models/CreateTaskRequestDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Task data
 */
export type CreateTaskRequestDTO = {
    courseId: string;
    academicTermYear: number;
    title: string;
    summary: string;
    dueDate: string;
    maxAttempts: number;
    minScoreToPass: number;
};