// src/app/core/api/models/TaskResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TaskResponseDTO = {
    id?: string;
    code?: string;
    title?: string;
    summary?: string;
    dueDate?: string;
    maxAttempts?: number;
    minScoreToPass?: number;
};