// src/app/core/api/models/UpdateTaskRequestDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Updated task data
 */
export type UpdateTaskRequestDTO = {
    title: string;
    summary: string;
    dueDate: string;
    maxAttempts: number;
    minScoreToPass: number;
};